move duplicated utility methods to utils module
